// ArchiCore/src/app/api-client/model/role.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Permission } from './permission';


export interface Role { 
    id?: number;
    name?: string;
    permissions?: Set<Permission>;
}